When we get a new window, check for size hints and compare them to the initial geometry. If they disagree, use the size hints instead.
Use 0,0 as origo! Duh.

If we get a new very large window, resize it to fit our screen before
mapping.

Use the XCB_EVENT_MASK_STRUCTURE_NOTIFY mask on the root window, so we
get CONFIGURE NOTIFY events from the root window itself! This means
that if RANDR suddenly changes the geometry we get to know about it.
Handle it gracefully.