Every client now remembers its place in each workspace window list. See struct client's wsitem[] array.
We use this to be able to free a window from a workspace window list
without having to traverse all items in all workspace lists.

This is stored as an array in the client because a window should be
able to occupy several workspaces simultaneously.

This also made it possible to clean up changing workspaces.

Also added early breaks in focusnext().