Always work on the currently focused window when moving and resizing and not in the window the mouse cursor happen to be in when the mouse button is pressed.
However, just do move and resize if the mouse cursor is inside the
window we care about. There's a limit to sloppy focus. It gets
confusing and we might lose focus to another window if we allow move
and resize *and continue to put the mouse cursor back to where it was
when starting, relative to the window*, as we do now.

Removed code for special handling of the root window. We don't use it
anyway.

Extra checks for focuswin == NULL.