Defined constant WORKSPACES for number of workspaces instead of WORKSPACE_MAX and used it consistently.
When fixing a window in fixwindow() we now add the window to all
workspace window lists except current in which it already should be
present. This means fixed windows are always a natural part of the
window list. No special treatment means that we don't muck around with
the order of the lists when changing workspaces.