arrangewindows() rewritten. We already know the geometry of windows since we track it all the time now.
Only call arrangewindows() when root window changes geometry if we
don't have RANDR. Otherwise, we call getrandr() when we get a
XCB_RANDR_SCREEN_CHANGE_NOTIFY event.

Added comments.

When a monitor gets a new mode in an RANDR scenario, we track the x,y
position and width,height.

If we lose a monitor, we add the windows on that monitor to another
monitor.