Use the X server's Save Set to store the windows we know about. This means they will be get automatically mapped and reparented if mcwm exits even if it crashes.
Simplified cleanup() to reflect the above change. We still need to
change the keyboard focus policy from mcwm's default to follow pointer
instead. If mcwm crashes completely before we reach cleanup() this
might become a problem. All the windows will be mapped but there won't
be a keyboard focus and no way to change it. I don't know how to fix
that right now.

Don't attempt to catch SIGSEGV. Might be dangerous.